Do not ask converter for files which are not required
Do not ask converter to output all file types (CAP, EXP and JCA). Make
it output only required files.

+++ src/pro/javacard/ant/JavaCard.java
@@ -466,7 +466,14 @@ public class JavaCard extends Task {
 			j.createArg().setLine("-verbose");
 			j.createArg().setLine("-nobanner");
 
-			j.createArg().setLine("-out CAP EXP JCA");
+			String outputs = "CAP";
+			if (output_exp != null) {
+				outputs += " EXP";
+			}
+			if (output_jca != null) {
+				outputs += " JCA";
+			}
+			j.createArg().setLine("-out " + outputs);
 			for (JCApplet app : raw_applets) {
 				j.createArg().setLine("-applet " + hexAID(app.aid) + " " + app.klass);
 			}
@@ -503,11 +510,10 @@ public class JavaCard extends Task {
 				java.nio.file.Path exp = jcsrc.resolve(ln + ".exp");
 				java.nio.file.Path jca = jcsrc.resolve(ln + ".jca");
 
-				if (!cap.toFile().exists() || !exp.toFile().exists() || !jca.toFile().exists()) {
-					throw new BuildException("Can not find CAP/EXP/JCA in " + jcsrc);
-				}
-
 				try {
+					if (!cap.toFile().exists()) {
+						throw new BuildException("Can not find CAP in " + jcsrc);
+					}
 					// Resolve output file
 					File opf = getProject().resolveFile(output_cap);
 					// Copy CAP
@@ -516,6 +522,9 @@ public class JavaCard extends Task {
 					// Copy exp file
 					if (output_exp != null) {
 						setTaskName("export");
+						if (!exp.toFile().exists()) {
+							throw new BuildException("Can not find EXP in " + jcsrc);
+						}
 						// output_exp is the folder name
 						opf = getProject().resolveFile(output_exp);
 
@@ -543,6 +552,9 @@ public class JavaCard extends Task {
 					// Copy JCA
 					if (output_jca != null) {
 						setTaskName("jca");
+						if (!jca.toFile().exists()) {
+							throw new BuildException("Can not find JCA in " + jcsrc);
+						}
 						opf = getProject().resolveFile(output_jca);
 						Files.copy(jca, opf.toPath(), StandardCopyOption.REPLACE_EXISTING);
 						log("JCA saved to " + opf.getAbsolutePath(), Project.MSG_INFO);
